merge from origin (#1)
* Update jepsen workflow. (#96)

* Update jepsen test readme. (#98)

This change reflects renamed arguments as well as the new redis-repo and
raft-repo arguments.

* Update raft library submodule (#101)

* Add and report version info/git sha. (#105)

* Add timeout to cluster joining (#103)

Before this commit, a RAFT.CLUSTER JOIN could loop forever.
Now, it is limited by the join-timeout configuration parameter.
Also, added connection-timeout to limit connection time for specific nodes.

* Tests: make fsync optional, fix timing issues. (#108)

* Fix crashes with large AE messages. (#109)

Before this commit, the Redis command was constructed on the stack which
could lead to overflow and crash. Now everything is heap based.

We should also address and optimize the root cause:
* Apply some limit on the number of entries per message
* Create a mechanism that would trigger a snapshot and deliver that
instead of a huge number of entries.

* Synchronous RAFT.CLUSTER JOIN (#110)

Before this commit, the command would return immediately and the
join operation would take place in the background.

Now, the join operation takes place immediately and the operation
completes only in one of these cases:

* Successfully executing `RAFT.NODE ADD` on the leader
* Timing on on `join-timeout`
* Receiving an unrecoverable error

Co-authored-by: Yossi Gottlieb <yossigo@gmail.com>

* Shardgroup fixes. (#106)

* Reduce logging verbosity of updates.
* Improve logging on multi-cluster tests.
* Add a test for shardgroup propagation.
* Fix shardgroup propagation.
* Fix an issue resulting with `<redisraft> {conn:6} Failed to resolve '': unknown node or service` errors by storing the last attempted address in `conn_addr`. If the connection is successful, `use_conn_addr` is set to `true` so a subsequent reconnect will first attempt to reconnect to the same address before iterating nodes.

Before this fix `conn_addr` would not be initialized. Upon successful connection, `use_conn_addr` was set to true and if a reconnection was required it would use the uninitialized `conn_addr`.

* unify joining and shard group link code (#112)

we now have a common set of connection handling code, instead of shared group linking and cluster joining having totally separate handling.

* Fix build issues with TRACE=1 (#115)

* Fix error message for RAFT.CONFIG SET loglevel trace (#116)

* Produce a more standard error on wrong args.

* Create a single common command spec and table.

* More environment variable friendliness.

All variables can now be set in advance making it possible to configure
the tool without a config.sh.

* Add tests/redis-suite/run.sh.

Provisions a local RedisRaft cluster using utils/create-cluster and
executes the Redis test suite against it in external mode.

Still work in progress - currently this mainly scopes what works and
what needs more refinement.

Co-authored-by: Yossi Gottlieb <yossigo@gmail.com>
Co-authored-by: Shaya Potter <spotter@gmail.com>
Co-authored-by: Dipankar Achinta <di.punk.car19@gmail.com>